Fixes bad bug that caused the parallel program to deadlock sometimes
If the number of simulations wasn't divisible with the number of
processes a deadlock resulted and the program never finished.

This seems to have been caused by a call to
ase.md.velocitydistribution.MaxwellBoltzmannDistribution which by
default makes a broadcast which could deadlock the program when all
processes didn't call this function the same number of times.