Temp solve ale multi bug, can run ale without installing it
This structure seems to cause problems with ale multi since it can't find
parallel_mpi_script.py when starting a new process with mpirun.
This was solved by adding the path to the parallel_mpu_script
when starting mpi process but could probably be solved in a better way
since this solution requires the user to execute ale multi from the
folder where the ale folder is located.

Adds instructions to README covering how to run ale without installing
it.

Running ale multi requires install with pip.